+++ CHANGELOG.md
@@ -46,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Another experimental mechanism to change 3D models according to the variant.
   (#103)
 - A mechanism to avoid running some outputs by default. (#112)
-- New pre-flight command to replace tags in the schematic. (#93)
+- New pre-flight commands to replace tags in the schematic and PCB. (#93)
 
 ### Changed
 - Internal BoM: now components with different Tolerance, Voltage, Current